Support identifier warehouses (#308)
* Support identifier warehouses

This is a bit confusing if you come from a Hive background
where the warehouse is always a path to hdfs/s3/etc.

With the REST catalog, the warehouse can also be a logical
identifier:
https://github.com/apache/iceberg/blob/main/open-api/rest-catalog-open-api.yaml#L72-L78

This means that we have to make sure that we only parse paths
that are an actual path, and not an identifier.

I'm open to suggestions. The check is now very simple, but can
be extended for example using a regex. But I'm not sure what
the implications are of importing additional packages (in Python
you want to keep it as lightweight as possible).

* Use `if Url::parse().is_ok()`