Retry object store reads on temporary errors.
I noticed that, when reading many parquet files from S3, one of the
reads fails occasionally with a temporary error such as "connection
closed before message completed". I think such transient failures are
normal for S3, and the client is expected to retry the read in this
case. However, opendal was not configured to perform a retry. This
commit adds a retry layer to opendal with the default configuration
(exponential backoff with up to 3 retries).

This solves the issue for me. It might be useful to make this behavior
configurable in the future, but default settings should work most of the
time.

Signed-off-by: Leonid Ryzhyk <leonid@feldera.com>