Support `{Positional,Equality}Deletes` (#652)
This PR adds support for handling of both positional and equality delete
files within table scans.

The approach taken is to include a list of delete file paths in every
`FileScanTask`. At the moment it is assumed that this list refers to
delete files that **may** apply to the data file in the scan, rather
than having been filtered to only contain delete files that definitely
do apply to this data file. Further optimisation of `plan_files` is
expected in the future to ensure that this list is pre-filtered before
being included in the FileScanTask.

This PR has been refactored so that this work can be split into multiple
PRs.

This PR now contains solely the scan plan changes.

---------

Co-authored-by: Fokko Driesprong <fokko@apache.org>