fix: Misleading error messages in `iceberg-catalog-rest` and allow `StatusCode::OK` in responses (#962)
Several of the error messages in `iceberg-catalog-rest` had misleading
error messages, so this PR fixes some of those.

Additionally, several of the `Catalog` methods on `RestCatalog` do not
accept `StatusCode::OK` as a successful response, so I've added that in
as well. This meant that some of the `client.rs` functions had to
change, but given that they are only used so many times I thought it was
fine to remove the const generics.

There are also a few other things that could be cleaned up but I'll keep
this PR short.