fix: TableMetadata `last_updated_ms` not increased for all operations (#978)
Currently we increase the `last_updated_ms` timestamp only if a snapshot
was added.
Java always updates this timestamp, also if i.e. only Properties where
added..

Trino has a catalog integration test that validates this - which we
failed due to this.

This PR ensures that `last_updated_ms` is always updated for builds.