feat: Add `SnapshotSummaries` (#1085)
## Which issue does this PR close?

- Related to #724 

## What changes are included in this PR?
This is the building block to implementing snapshot summaries. Most of
the implementation code is built off of @barronw 's very nice commit.
Most of the added lines are also just tests. A follow up pr will be
included integrating it with the rest of the codebase.

## Are these changes tested?
Yes, unit test