Remove `paste` dependency by expanding previously macro-generated code (#1138)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #1064.

## What changes are included in this PR?

<!--
Provide a summary of the modifications in this PR. List the main changes
such as new features, bug fixes, refactoring, or any other updates.
-->

* Remove `paste` as a dependency from Cargo files
* Expand code that was previously generated through macros that used
`paste!` (as suggested in #1064)

## Are these changes tested?

<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

Given that this is a refactoring that does not add new functionality, I
rely on the existing test suite to surface any errors.