Scan Delete Support Part 3: `ArrowReader::build_deletes_row_selection` implementation (#951)
Third part of delete file read support. See
https://github.com/apache/iceberg-rust/issues/630

**Builds on top of https://github.com/apache/iceberg-rust/pull/950


`build_deletes_row_selection` computes a `RowSelection` from a
`RoaringTreemap` representing the indexes of rows in a data file that
have been marked as deleted by positional delete files that apply to the
data file being read (and, in the future, delete vectors).

The resulting `RowSelection` will be merged with a `RowSelection`
resulting from the scan's filter predicate (if present) and supplied to
the `ParquetRecordBatchStreamBuilder` so that deleted rows are omitted
from the `RecordBatchStream` returned by the reader.

NB: I encountered quite a few edge cases in this method and the logic is
quite complex. There is a good chance that a keen-eyed reviewer would be
able to conceive of an edge-case that I haven't covered.

---------

Co-authored-by: Renjie Liu <liurenjie2008@gmail.com>