refactor: Bump MSRV to 1.84 for preparing next release (#1185)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

None

## What changes are included in this PR?

Bump MSRV to 1.84 to get our community ready for the next release.

Rust 1.84 is released on 9 January, 2025 which is 3 months ago.

We need this PR in advance since we still need time to play well with
rust 1.84 like using new deps or better APIs and so so.

## Are these changes tested?

<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

---------

Signed-off-by: Xuanwo <github@xuanwo.io>