feat: Allow reuse http client in rest catalog (#1221)
## Which issue does this PR close?

Creating a reqwest client is not without cost; it involves DNS caching,
HTTP connection pooling, and TLS certificate loading. The previous
design also made it difficult for users to configure the HTTP client,
such as setting custom timeout values or using a globally shared DNS
resolver.

This PR addressed this by allowing users to create the client and pass
to rest catalog as config.

## What changes are included in this PR?

Allow users to pass a `reqwest::Client` as config.

## Are these changes tested?

Unit tests

Signed-off-by: Xuanwo <github@xuanwo.io>