fix(catalog/rest): Using async lock in token to avoid blocking runtime (#1223)
## Which issue does this PR close?

We used to think it was acceptable to use a blocking lock in the token
since we were not crossing an await boundary. However, our users
reported that this can cause the runtime to hang completely if multiple
catalog instances try to acquire the token concurrently.

## What changes are included in this PR?

This PR fixed it by using an async lock instead.

## Are these changes tested?

Unit tests.

Signed-off-by: Xuanwo <github@xuanwo.io>
Co-authored-by: Renjie Liu <liurenjie2008@gmail.com>