fix: Update view version log timestamp for historical accuracy (#1218)
## What changes are included in this PR?

This change updates the behavior when setting a view version as current.
When we set a previously existing view version as current, we now update
its log timestamp to the current time rather than reusing the original
timestamp of the ViewVersion.

Lets assume the following changes:
1. ViewVersion 1 is added and set active
2. View Version 2 is added and set active
3. View Version 1 is set active.

This resulted in the following `version_log`:
```json
    "version-log": [
        {
            "version-id": 1,
            "timestamp-ms": 1744716416168
        },
        {
            "version-id": 2,
            "timestamp-ms": 1744716449754
        },
        {
            "version-id": 1,
            "timestamp-ms": 1744716416168
        }
    ],
```

Note that the last and first entries have the same timestamp, namely the
time stamp of the initial ViewVersion creation.

I believe this is undesired in a history, where we are interested when a
certain change became active. It makes sense to use the exact timestamp
of the ViewVersion if it was added in the same set of changes, but
re-enabling a previously used view version (maybe years ago) should not
add a history for this past timestamp.

Java behaviors the same way as rust currently does. @Fokko @nastra if we
agree that we should change the behavior, we should also touch Java.

## Are these changes tested?

Yes