feat: re-export name mapping (#1116)
## Which issue does this PR close?

Likely helps towards https://github.com/apache/iceberg-rust/issues/919
and this was also discussed in
[Slack](https://apache-iceberg.slack.com/archives/C05HTENMJG4/p1742385647684059).

## What changes are included in this PR?

This publicly re-exports the `name_mapping` module to `iceberg::spec`.
Prior to this, it is private and inaccessible outside of this crate.

## Are these changes tested?

The main changes here are not functional changes except to visibility.

The additional fields to the `NameMapping` structure have test coverage
and are covered by prior tests, it is expected that these internal
fields do **not** show up in the serialised `NameMapping` output, hence
they are marked with the `serde(skip)` attribute - this is upheld with
prior tests.

---------

Co-authored-by: Renjie Liu <liurenjie2008@gmail.com>