refactor(s3tables): avoid misleading FileIO::from_path
The `FileIO::from_path` only validates a s3 url, and finally calls `FileIOBuilder::new(url.scheme())`.

For s3 tables, the table bucket is not a real bucket, and we can't do s3 operations on that. The `FileIO::from_path` here gives a fake impression that the bucket name will be used.

Signed-off-by: xxchan <xxchan22f@gmail.com>