feat: support arrow dictionary in schema conversion (#1293)
## Which issue does this PR close?


- Closes https://github.com/apache/iceberg-rust/issues/1277

## What changes are included in this PR?

An expansion to the current schema conversion for arrow to an iceberg
schema.

This uses the suggestion provided by @alamb
([here](https://github.com/apache/iceberg-rust/issues/1277#issuecomment-2845952969))
for the schema conversion work.


## Are these changes tested?

The current test for schema conversion has been expanded.

I've also taken the liberty of altering the `assert_eq!` call to use the
`pretty_assertions` version, as it makes viewing the failure much
simpler with a large JSON structure - I can revert this though if it is
problematic.