ci: Fix python bindings rust code not checked (#1338)
## Which issue does this PR close?

None

## What changes are included in this PR?

While reviewing PRs related to the python bindings, I noticed that the
code formatting was incorrect and some clippy lints were not being
triggered properly.

This happened because we weren't running the checks on the correct path.
This PR addresses that issue.

Our Python friends should be happier and feel more at home with Rust
now.

## Are these changes tested?

Not needed, it's in CI.

Signed-off-by: Xuanwo <github@xuanwo.io>