Change `FieldSummary` `{upper,lower}_bound` to `ByteBuf` (#1369)
## Which issue does this PR close?

I would like to invite everyone to roast my Rust-skills in order for me
to improve myself :)

Unblocks https://github.com/apache/iceberg-rust/pull/1328

This aligns closely with PyIceberg and Java and greatly simplifies the
use of the Avro readers in PyIceberg. Otherwise, we would need to update
public APIs.

## What changes are included in this PR?



## Are these changes tested?



---------

Co-authored-by: Kevin Liu <kevinjqliu@users.noreply.github.com>