refine: refine the interface of SnapshotProducer (#1490)
## Which issue does this PR close?


## What changes are included in this PR?


This PR refine the interface of SnapshotProducer:
1. include the table in SnapshotProducer rather than pass using function
param
2. add SnapshotProducer as param for ManifestProcess,
SnapshotProduceOperation

I find that it would convenient when working on merge append:
ManifestProcess and SnapshotProduceOperation can be consider as custom
extension for SnapshotProducer and they would reuse SnapshotProducer as
for common usage. So we should include it in their function param. At
the same time, include Table in SnapshotProducer make thing easier they
can use SnapshotProducer diretcly.

## Are these changes tested?


---------

Co-authored-by: ZENOTME <st810918843@gmail.com>
Co-authored-by: Renjie Liu <liurenjie2008@gmail.com>