fix: Fix mock dependency (#1507)
## What changes are included in this PR?

`mockall` dependency is introduced in
https://github.com/apache/iceberg-rust/pull/1484, which is used for unit
test on retry.
But it's only used in unit tests, but not production code; so it should
be placed under dev section, and completely disabled in production code
path.
Current implementation unnecessarily increases crate size for production
usage.

## Are these changes tested?

N/A, a no-op change, a passed compilation should be enough