Add support for custom credential loader for S3 FileIO (#1528)
## Which issue does this PR close?

- Closes #1527

## What changes are included in this PR?

Adds the ability to provide custom extensions to the `FileIOBuilder`.
Currently the only supported extension is `CustomAwsCredentialLoader`
which is a newtype around
[`AwsCredentialLoad`](https://docs.rs/reqsign/0.16.3/reqsign/trait.AwsCredentialLoad.html),
which is what OpenDAL expects.

I've added extensions to the `RestCatalog` as well, and when its
constructing `FileIO` for table operations, passes along any defined
extensions into the `FileIOBuilder`, which then get passed into the
underlying OpenDAL constructors.

## Are these changes tested?

Yes, tests added in `crates/iceberg/tests/file_io_s3_test.rs` that
verify the extension is working.