Add LICENSE and NOTICE files for pyiceberg-core source distribution and wheel (#1653)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- None.

## What changes are included in this PR?

This adds LICENSE and NOTICE files to the PyPI sdist which is needed for
distributing conda package. Also add those files to the wheel
distribution

<!--
Provide a summary of the modifications in this PR. List the main changes
such as new features, bug fixes, refactoring, or any other updates.
-->

## Are these changes tested?

<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

---------

Co-authored-by: Kevin Liu <kevinjqliu@users.noreply.github.com>