Bump Avro to 0.20 (#1644)
## Which issue does this PR close?

Because of improved error messages:
https://github.com/apache/avro-rs/pull/228. This helps a lot when
working on using the Iceberg Avro reader in PyIceberg. Also, when we
ship the Avro reader with PyIceberg, it will allow users for creating
meaningful issues.

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #.

## What changes are included in this PR?

<!--
Provide a summary of the modifications in this PR. List the main changes
such as new features, bug fixes, refactoring, or any other updates.
-->

## Are these changes tested?

<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->