refactor: Move equality-ids closer to the spec (#1705)
## Which issue does this PR close?

Right now the equality-deletes can't be not-null, while the spec states
that it should be null in the case of a manifest-entry that's not an
equality delete:

<img width="835" height="341" alt="image"
src="https://github.com/user-attachments/assets/60a88f37-7c50-48b7-8878-ecfe4bd70509"
/>


## What changes are included in this PR?

<!--
Provide a summary of the modifications in this PR. List the main changes
such as new features, bug fixes, refactoring, or any other updates.
-->

## Are these changes tested?

<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->