fix: global eq delete matching should apply to only strictly older files, and fix partition scoped matching to consider spec id (#1758)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #1759.

## What changes are included in this PR?

This changes scan planning of equality deletes so that: 
1. we only match against eq deletes which are strictly older than the
data file. Right now it looks like we incorrectly over-apply based on
the seq number for global equality deletes.

2. Partition scoped deletes (both equality and position) are compared
correctly by also factoring in the spec ID. It's not quite enough to
compare just off the tuple, we should also compare based off the spec ID
as well.

## Are these changes tested?

Added unit tests which are scoped to testing delete index matching
logic.

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->