feat(reader): Add Date32 support to RecordBatchTransformer create_column (#1792)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Partially address #1749.

Iceberg Java has a test that does a schema change requires the
`RecordBatchTransformer` to add a Date32 column, which it currently does
not support.

## What changes are included in this PR?

<!--
Provide a summary of the modifications in this PR. List the main changes
such as new features, bug fixes, refactoring, or any other updates.
-->

Add match arms for `Date32` type in `create_column`.

## Are these changes tested?

<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

New test that mirrors Iceberg Java's
`TestSelect.readAndWriteWithBranchAfterSchemaChange`

---------

Co-authored-by: Xuanwo <github@xuanwo.io>