ci: Don't dismiss stale review to make contribution easier (#1799)
## Which issue does this PR close?


- Closes #.

## What changes are included in this PR?

I made this change to avoid dismissing stale reviews to make
contributions easier.

iceberg is quite active, and we nearly have conflicts every day. It can
be a pain for contributors to get approved, especially after resolving
conflicts—they then have to wait for reviewers to give another approval.

I think we can make it clear that trivial changes like resolving merge
conflicts are allowed, but adding other code changes aren't. We can just
add this to the docs so it’s easier for everyone.

## Are these changes tested?