infra: add docs for cleaning up testpypi artifacts (#1855)
## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

- Closes #1837

## What changes are included in this PR?
Used the [`pypi-cleanup`](https://github.com/arcivanov/pypi-cleanup)
utility to bulk cleanup old artifacts of pyiceberg-core from testpypi
for #1837. The alternative is to do this manually in the UI for each
artifact.

Adding this as reminder / instruction when the nightly pipeline fails
again

<!--
Provide a summary of the modifications in this PR. List the main changes
such as new features, bug fixes, refactoring, or any other updates.
-->

## Are these changes tested?

<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

Co-authored-by: Fokko Driesprong <fokko@apache.org>