chore: bump MSRV to 1.88, fix warnings and clippy errors (#1902)
https://github.com/apache/iceberg-rust/pull/1899 requires a bump to MSRV
1.88. This version is within the policy of this project, and since the
README mentions `MSRV is updated when we release iceberg-rust` and we're
preparing 0.8, here's a PR for just MSRV 1.88.

## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and
enhancements and this helps us generate change logs for our releases.
You can link an issue to this PR using the GitHub syntax. For example
`Closes #123` indicates that this PR will close issue #123.
-->

N/A.

## What changes are included in this PR?

<!--
Provide a summary of the modifications in this PR. List the main changes
such as new features, bug fixes, refactoring, or any other updates.
-->

- Bump MSRV to 1.88
- Fix warnings
- Fix errors found by `make check-clippy`
- Format

## Are these changes tested?

<!--
Specify what test covers (unit test, integration test, etc.).

If tests are not included in your PR, please explain why (for example,
are they covered by existing tests)?
-->

Existing tests